Move spy connections to dedicated namespace
