+++ tests/functional/MysqliTest.php
@@ -6,6 +6,7 @@ namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest;
 
 use Doctrine\DBAL\Driver\Mysqli\Driver;
 use Doctrine\DBAL\DriverManager;
+use Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest\Spy\Connection;
 
 class MysqliTest extends AbstractFunctionalTestCase
 {
