+++ tests/functional/PDOMySqlTest.php
@@ -6,6 +6,7 @@ namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest;
 
 use Doctrine\DBAL\Driver\PDO\MySQL\Driver;
 use Doctrine\DBAL\DriverManager;
+use Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest\Spy\Connection;
 
 class PDOMySqlTest extends AbstractFunctionalTestCase
 {
