+++ tests/functional/Spy/Connection.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest;
+namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest\Spy;
 
 /**
  * @psalm-suppress PropertyNotSetInConstructor
@@ -11,6 +11,9 @@ class Connection extends \Facile\DoctrineMySQLComeBack\Doctrine\DBAL\Connection
 {
     public int $connectCount = 0;
 
+    /**
+     * @param string|null $connectionName
+     */
     public function connect($connectionName = null): bool
     {
         if (! $this->isConnected()) {
