+++ tests/functional/Spy/PrimaryReadReplicaConnection.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest;
+namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL\FunctionalTest\Spy;
 
 /**
  * @psalm-suppress PropertyNotSetInConstructor
